finished promotions

+++ app-order/src/main/java/uz/pdp/apporder/payload/promotion/BonusProductPromotionDTO.java
@@ -16,7 +16,7 @@ public class BonusProductPromotionDTO {
 
     private Float moreThan;
 
-    private Float bonusCount;
+    private Short bonusCount;
 
     private ProductDTO productDTO;
 }
